refactor(pipeline): split into pipeline/ module (orchestration vs layers)
Separate concerns so adding a layer is a new file, not bloat in one:
- pipeline/mod.rs   — Pipeline, Layers, run()/stream() (layer-agnostic)
- pipeline/levels.rs — Levels + cached env/config resolution
- pipeline/decorative.rs — the decorative layer + its streaming adapter

No behaviour change; pure code movement.
